REFACTORED code in coordinatorService to optimize code

+++ utility/index.js
@@ -1,10 +1,11 @@
-const jwt               = require('jsonwebtoken'),
-      message           = "Something went wrong, try again",
-      fs                = require('fs'),
-      { upload }        = require('../config/multer'),
-      config            = require('../config').get(),
-      { s3ImagesLocal } = config.awsCredentials,
-      unauthorize       = { success: false, statuscode: 401, message: "Unauthorized user" };
+const jwt = require('jsonwebtoken'),
+    message = "Something went wrong, try again",
+    fs = require('fs'),
+    model = require('../app/model'),
+    { upload } = require('../config/multer'),
+    config = require('../config').get(),
+    { s3ImagesLocal } = config.awsCredentials,
+    unauthorize = { success: false, statuscode: 401, message: "Unauthorized user" };
 
 module.exports = {
     generateToken: (payload) => {
@@ -31,9 +32,9 @@ module.exports = {
         }
         catch (err) {
             return res.status(401).json({
-                auth      : false,
+                auth: false,
                 statuscode: 401,
-                message   : 'Failed to authenticate'
+                message: 'Failed to authenticate'
             });
         }
     },
@@ -55,7 +56,40 @@ module.exports = {
         })
     },
 
-    deleteImg: (file) => {
+    deleteImg: deleteImg,
+
+    checkUserAlreadyExist: (req, res, next) => {
+        let queryObj = [];
+        let keys = Object.keys(req.body);
+        for (var i = 0; i < keys.length; i++) {
+            let obj = {};
+            if (["mobile", "emailId"].includes(keys[i])) {
+                obj[keys[i]] = req.body[keys[i]];
+                queryObj.push(obj);
+            }
+            continue;
+        }
+        if (req.body.mobile || req.body.emailId) {
+            model.findCoordinator({ $or: queryObj }, (err, data) => {
+                if (err) {
+                    logger.error("error occur in checkUserAlreadyExist middleaware find query callback")
+                    return res.status(400).send({ message: "Something went wrong", statuscode: 400 })
+                }
+                if (data.length !== 0) {
+                    let val = data[0].mobile === Number(req.body.mobile) ? req.body.mobile : req.body.emailId;
+                    [undefined, "", null].includes(req.file) ? null : deleteImg(req.file);
+                    return res.status(409).send({
+                        message: `[${val}] already exist`,
+                        statuscode: 409
+                    })
+                } else return next()
+            })
+        } else return next()
+    }
+}
+
+function deleteImg(file) {    
+    if (![undefined, "", null].includes(file)) {
         fs.unlink(`./${s3ImagesLocal}/${file}`, (err) => {
             if (err) logger.error("Failed to delete local image:" + err);
             else logger.info('Successfully deleted local image');
